Finish reducing backend duplication on bot routes, use botExists middleware.
